--- diagrams/example.drawio
@@ -1,40 +1,0 @@
-<mxfile host="65bd71144e">
-    <diagram id="jxCgwqO4SBPk6OXGrDTB" name="Page-1">
-        <mxGraphModel dx="1142" dy="901" grid="0" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="413" pageHeight="583" math="0" shadow="0">
-            <root>
-                <mxCell id="0"/>
-                <mxCell id="1" parent="0"/>
-                <mxCell id="2" value="Add a diagram file in the repo_root/diagrams/ dir named &quot;diagram_name.drawio&quot;" style="rounded=0;whiteSpace=wrap;html=1;" parent="1" vertex="1">
-                    <mxGeometry x="120" y="160" width="160" height="80" as="geometry"/>
-                </mxCell>
-                <mxCell id="3" value="Install the &quot;draw.io integration&quot; extension" style="rounded=0;whiteSpace=wrap;html=1;" parent="1" vertex="1">
-                    <mxGeometry x="120" y="80" width="160" height="40" as="geometry"/>
-                </mxCell>
-                <mxCell id="4" value="VSCode will use the extension to open the diagram in a draw.io window" style="rounded=0;whiteSpace=wrap;html=1;" parent="1" vertex="1">
-                    <mxGeometry x="120" y="280" width="160" height="80" as="geometry"/>
-                </mxCell>
-                <mxCell id="5" value="Save the file, the tool will automatically generate .png, .svg, and .pdf diagrams." style="rounded=0;whiteSpace=wrap;html=1;" parent="1" vertex="1">
-                    <mxGeometry x="120" y="400" width="160" height="80" as="geometry"/>
-                </mxCell>
-                <mxCell id="6" value="" style="endArrow=classic;html=1;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" parent="1" source="3" target="2" edge="1">
-                    <mxGeometry width="50" height="50" relative="1" as="geometry">
-                        <mxPoint x="170" y="350" as="sourcePoint"/>
-                        <mxPoint x="220" y="300" as="targetPoint"/>
-                    </mxGeometry>
-                </mxCell>
-                <mxCell id="7" value="" style="endArrow=classic;html=1;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" parent="1" source="2" target="4" edge="1">
-                    <mxGeometry width="50" height="50" relative="1" as="geometry">
-                        <mxPoint x="210" y="130" as="sourcePoint"/>
-                        <mxPoint x="210" y="170" as="targetPoint"/>
-                    </mxGeometry>
-                </mxCell>
-                <mxCell id="8" value="" style="endArrow=classic;html=1;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" parent="1" source="4" target="5" edge="1">
-                    <mxGeometry width="50" height="50" relative="1" as="geometry">
-                        <mxPoint x="220" y="140" as="sourcePoint"/>
-                        <mxPoint x="220" y="180" as="targetPoint"/>
-                    </mxGeometry>
-                </mxCell>
-            </root>
-        </mxGraphModel>
-    </diagram>
-</mxfile>
