+++ src/daqpyutils/__init__.py
@@ -1,0 +1,3 @@
+from daqpytools.logging.logger import setup_root_logger
+
+setup_root_logger(logger_name="daqpyutils", log_level="ERROR")
