+++ src/daqpyutils/repository_handling/__init__.py
@@ -1,0 +1,3 @@
+from daqpytools.logging.logger import get_daq_logger
+
+get_daq_logger(logger_name="daqpyutils.repository_handling")
