+++ src/daqpyutils/templates/pyproject.jinja
@@ -58,7 +58,7 @@ omit = ["tests/*"]
 {% if STRICT -%}
 [tool.ruff]
 line-length = 88
-{%- endif -%}
+{%- endif %}
 
 [tool.ruff.lint]
 ignore = ["D205", "D100", "D104"]
